improved /schematics list total filesize output need to add automatic conversion into Mb in the future

+++ worldedit-core/src/main/java/com/boydti/fawe/config/Settings.java
@@ -399,7 +399,7 @@ public class Settings extends Config {
                 "Sets a maximum limit (in kb) for the size of a player's schematics directory (per-player mode only)",
                 "set to -1 to have no limit"
         })
-        public double PERPLAYER_FILESIZELIMIT = -1;
+        public int PERPLAYER_FILESIZELIMIT = -1;
 
         @Comment({
                 "Sets a maximum limit for the amount of schematics in a player's schematics directory (per-player mode only)",
